ec2-batch-processor-gpu/requirements.txt
boto3>=1.34.0
numpy>=1.24.0
opencv-python-headless>=4.5.0
Pillow>=10.0.0
paddlepaddle-gpu==3.0.0
paddleocr==2.9.1
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
psycopg[binary]>=3.1.0
pgvector>=0.2.0
python-dotenv>=1.0.0
